refactor: move help article object to utils constants
