+++ apps/web/src/components/balances/TokenListSelect/index.tsx
@@ -7,9 +7,9 @@ import ExternalLink from '@/components/common/ExternalLink'
 import { OnboardingTooltip } from '@/components/common/OnboardingTooltip'
 import Track from '@/components/common/Track'
 import { ASSETS_EVENTS, trackEvent } from '@/services/analytics'
-import { HelpCenterArticle } from '@/config/constants'
 import { useHasFeature } from '@/hooks/useChains'
 import { FEATURES } from '@safe-global/utils/utils/chains'
+import { HelpCenterArticle } from '@safe-global/utils/config/constants'
 
 const LS_TOKENLIST_ONBOARDING = 'tokenlist_onboarding'
