+++ apps/web/src/components/common/ErrorBoundary/index.tsx
@@ -1,12 +1,13 @@
 import { Typography, Link } from '@mui/material'
 
-import { HELP_CENTER_URL, IS_PRODUCTION } from '@/config/constants'
+import { IS_PRODUCTION } from '@/config/constants'
 import { AppRoutes } from '@/config/routes'
 import WarningIcon from '@/public/images/notifications/warning.svg'
 
 import css from '@/components/common/ErrorBoundary/styles.module.css'
 import CircularIcon from '../icons/CircularIcon'
 import ExternalLink from '../ExternalLink'
+import { HELP_CENTER_URL } from '@safe-global/utils/config/constants'
 interface ErrorBoundaryProps {
   error: Error
   componentStack: string
