+++ apps/web/src/components/safe-apps/SafeAppsErrorBoundary/SafeAppsLoadError.tsx
@@ -1,11 +1,11 @@
 import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
 import SvgIcon from '@mui/material/SvgIcon'
-import { DISCORD_URL } from '@/config/constants'
 import NetworkError from '@/public/images/apps/network-error.svg'
 
 import css from './styles.module.css'
 import ExternalLink from '@/components/common/ExternalLink'
+import { DISCORD_URL } from '@safe-global/utils/config/constants'
 
 type SafeAppsLoadErrorProps = {
   onBackToApps: () => void
