+++ apps/web/src/components/settings/TransactionGuards/index.tsx
@@ -6,12 +6,12 @@ import css from './styles.module.css'
 import ExternalLink from '@/components/common/ExternalLink'
 import { SAFE_FEATURES } from '@safe-global/protocol-kit/dist/src/utils/safeVersions'
 import { hasSafeFeature } from '@/utils/safe-versions'
-import { HelpCenterArticle } from '@/config/constants'
 import DeleteIcon from '@/public/images/common/delete.svg'
 import CheckWallet from '@/components/common/CheckWallet'
 import { useContext } from 'react'
 import { TxModalContext } from '@/components/tx-flow'
 import { RemoveGuardFlow } from '@/components/tx-flow/flows'
+import { HelpCenterArticle } from '@safe-global/utils/config/constants'
 
 const NoTransactionGuard = () => {
   return (
