+++ apps/web/src/components/transactions/TxDetails/TxData/Rejection/index.tsx
@@ -3,7 +3,8 @@ import { NOT_AVAILABLE } from '@/components/transactions/TxDetails'
 import type { MultisigExecutionDetails } from '@safe-global/safe-gateway-typescript-sdk'
 import { Box, Typography } from '@mui/material'
 import React from 'react'
-import { HelpCenterArticle } from '@/config/constants'
+
+import { HelpCenterArticle } from '@safe-global/utils/config/constants'
 
 interface Props {
   nonce?: MultisigExecutionDetails['nonce']
